requirement.txt
altgraph==0.17
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
BTrees==4.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cx-Oracle==6.4.1
dis3==0.1.3
Django==1.11.16
django-allauth==0.38.0
django-appconf==1.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-imagekit==4.0.2
django-uuslug==1.1.8
docopt==0.6.2
EasyTkinter==1.1.0
enum34==1.1.6
ExtensionClass==4.4
future==0.18.2
futures==3.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
itchat==1.3.10
lxml==4.3.0
macholib==1.14
mod-wsgi==4.6.5+ap24vc9
oauthlib==2.1.0
olefile==0.44
pefile==2019.4.18
persistent==4.4.3
pilkit==2.0
Pillow==4.2.1
pycparser==2.19
pycryptodomex==3.9.7
pygame==1.9.3
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pypng==0.0.19
PyQRCode==1.2.1
python-gettext==4.0
python-openid==2.2.5
python-qt5==0.1.10
python-slugify==1.2.4
pytz==2017.2
pywin32-ctypes==0.2.0
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
soaplib==1.0.0
sorl-thumbnail==12.4a1
sqlparse==0.2.4
suds==0.4
termcolor==1.1.0
transaction==2.4.0
Unidecode==0.4.21
urllib3==1.23
zc.lockfile==1.4
ZConfig==3.4.0
zdaemon==4.3
zExceptions==4.1
ZODB==5.5.1
zodbpickle==1.0.3
zope.browser==2.3
zope.cachedescriptors==4.3.1
zope.component==4.5
zope.configuration==4.3.0
zope.container==4.2.2
zope.contenttype==4.4
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.dottedname==4.3
zope.event==4.4
zope.exceptions==4.3
zope.filerepresentation==4.2.0
zope.hookable==4.2.0
zope.i18n==4.6.1
zope.i18nmessageid==4.3.1
zope.interface==4.6.0
zope.lifecycleevent==4.3
zope.location==4.2
zope.proxy==4.3.1
zope.publisher==5.0.1
zope.schema==4.9.3
zope.security==4.3.1
zope.size==4.3
zope.traversing==4.3.1
